Change Test failed/passed to Longs to work around server class cast bug.

+++ src/foam/nanos/test/Test.js
@@ -29,7 +29,7 @@ foam.CLASS({
   properties: [
     'id',
     {
-      class: 'Int',
+      class: 'Long',
       name: 'passed',
       visibility: foam.u2.Visibility.RO,
       tableCellFormatter: function(value) {
@@ -37,7 +37,7 @@ foam.CLASS({
       }
     },
     {
-      class: 'Int',
+      class: 'Long',
       name: 'failed',
       visibility: foam.u2.Visibility.RO,
       tableCellFormatter: function(value) {
